The most significant changes involve the addition of a new method and a new data structure to handle fake data in the `SnakeskinEntityFrameworkCoreExtensions.cs` file. The `CommitFake` method has been added to the `ModelBuilder` class to commit fake data for all types not in the `IgnoredTypes` list. The `IgnoredTypes` is a new `ConcurrentBag<Type>` that stores types to be ignored when faking data. Additionally, the `CommitFakeAsync` method has been marked as obsolete due to poor performance.
In the `SnakeskinDbContext.cs` file, a call to `modelBuilder.CommitFake();` has been added in the `OnModelCreating` method, indicating the use of the new `CommitFake` method.

In `Program.cs`, the line `await app.CommitFakeAsync<SnakeskinDbContext>();` has been commented out, suggesting that the asynchronous method to commit fake data to the `SnakeskinDbContext` is no longer in use.

Lastly, in `SnakeskinFakeTypeBuilder.cs`, a new method `GetTypes` has been added to the `SnakeskinFakeTypeCollection` class. This method returns all types in the collection that are not in the provided `ignoreTypes` list.

List of changes:

1. Addition of `CommitFake` method to `ModelBuilder` class in `SnakeskinEntityFrameworkCoreExtensions.cs` file.
2. Addition of `IgnoredTypes` data structure in `SnakeskinEntityFrameworkCoreExtensions.cs` file.
3. Marking `CommitFakeAsync` method as obsolete in `SnakeskinEntityFrameworkCoreExtensions.cs` file.
4. Addition of `modelBuilder.CommitFake();` call in `OnModelCreating` method in `SnakeskinDbContext.cs` file.
5. Commenting out of `await app.CommitFakeAsync<SnakeskinDbContext>();` line in `Program.cs` file.
6. Addition of `GetTypes` method to `SnakeskinFakeTypeCollection` class in `SnakeskinFakeTypeBuilder.cs` file.